refactor(api): merge routes into app.py, gemini model changed and update README
