+++ agents/app.py
@@ -1,12 +1,12 @@
-from fastapi import FastAPI
+from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
-from agents.routes.recommend_route import router as recommend_router
 from agents.utils.response_models import (
     SuggestionRequest, SuggestionResponse, WorkflowSuggestionResponse
 )
 from agents.suggesting_agent import ToolSuggestionAgent
 from agents.workflow_suggestion_agent import WorkflowSuggestionAgent
 from agents.summary_agent import summarize_tool_suggestions, summarize_workflow_suggestions
+from agents.classification_service import classify_query
 
 app = FastAPI(title="Galaxy Tool Suggestion API")
 
@@ -60,5 +60,35 @@ def suggest_workflows_enhanced(request: SuggestionRequest):
     )
     return {"results": results}
 
-# ------------------ GEMINI-POWERED ENDPOINT ----- #
-app.include_router(recommend_router)
+
+# ------------------ UNIFIED RECOMMENDATION ENDPOINT ----- #
+@app.post("/recommend")
+async def recommend(request: SuggestionRequest):
+    """
+    Unified endpoint:
+    - Gemini classifies the query as 'tool', 'workflow', or 'both'
+    - Routes to the correct agent(s)
+    - Returns structured JSON
+    """
+    try:
+        # Step 1: Gemini classification
+        category = classify_query(request.query)
+
+        # Step 2: Route to appropriate agent(s)
+        if category == "tool":
+            tools = agent.suggest_tools(request.query, request.top_k)
+            return {"type": "tool", "results": tools}
+
+        elif category == "workflow":
+            workflows = workflow_agent.suggest_workflows(request.query, request.top_k)
+            return {"type": "workflow", "results": workflows}
+
+        else:  # "both"
+            tools = agent.suggest_tools(request.query, request.top_k)
+            workflows = workflow_agent.suggest_workflows(request.query, request.top_k)
+            return {"type": "both", "tool_results": tools, "workflow_results": workflows}
+
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+
